Misc. fixes: - Missed Microsoft.Web.Infrastructure DLL in build output - Corrected/reversed sort order for revisions - Changed repo disk size field from int to long (for large repos) - Changed chart types - Additional favicon reference - Cleaned up web.config - Updated screenshot

+++ Weald/Models/RepositoryInfo.cs
@@ -8,7 +8,7 @@ namespace Weald.Models
         public int LatestRevision;
         public string LatestChangeUsername;
         public DateTime LatestChangeTimestamp;
-        public int SizeInBytes;
+        public long SizeInBytes;
         public string Url;
     }
 }
